+++ src/main/java/com/arangodb/internal/InternalArangoVertexCollection.java
@@ -46,6 +46,8 @@ public abstract class InternalArangoVertexCollection<A extends InternalArangoDB<
 	private static final String PATH_API_GHARIAL = "/_api/gharial";
 	private static final String VERTEX = "vertex";
 
+    private static final String TRANSACTION_ID = "x-arango-trx-id";
+
 	private final G graph;
 	private final String name;
 
@@ -71,6 +73,7 @@ public abstract class InternalArangoVertexCollection<A extends InternalArangoDB<
 		final Request request = request(graph.db().name(), RequestType.POST, PATH_API_GHARIAL, graph.name(), VERTEX,
 			name);
 		final VertexCreateOptions params = (options != null ? options : new VertexCreateOptions());
+        request.putHeaderParam(TRANSACTION_ID, params.getStreamTransactionId());
 		request.putQueryParam(ArangoRequestParam.WAIT_FOR_SYNC, params.getWaitForSync());
 		request.setBody(util(Serializer.CUSTOM).serialize(value));
 		return request;
@@ -93,6 +96,7 @@ public abstract class InternalArangoVertexCollection<A extends InternalArangoDB<
 		final Request request = request(graph.db().name(), RequestType.GET, PATH_API_GHARIAL, graph.name(), VERTEX,
 			DocumentUtil.createDocumentHandle(name, key));
 		final GraphDocumentReadOptions params = (options != null ? options : new GraphDocumentReadOptions());
+        request.putHeaderParam(TRANSACTION_ID, params.getStreamTransactionId());
 		request.putHeaderParam(ArangoRequestParam.IF_NONE_MATCH, params.getIfNoneMatch());
 		request.putHeaderParam(ArangoRequestParam.IF_MATCH, params.getIfMatch());
 		if (params.getAllowDirtyRead() == Boolean.TRUE) {
@@ -109,6 +113,7 @@ public abstract class InternalArangoVertexCollection<A extends InternalArangoDB<
 		final Request request = request(graph.db().name(), RequestType.PUT, PATH_API_GHARIAL, graph.name(), VERTEX,
 			DocumentUtil.createDocumentHandle(name, key));
 		final VertexReplaceOptions params = (options != null ? options : new VertexReplaceOptions());
+		request.putHeaderParam(TRANSACTION_ID, params.getStreamTransactionId());
 		request.putQueryParam(ArangoRequestParam.WAIT_FOR_SYNC, params.getWaitForSync());
 		request.putHeaderParam(ArangoRequestParam.IF_MATCH, params.getIfMatch());
 		request.setBody(util(Serializer.CUSTOM).serialize(value));
@@ -131,6 +136,7 @@ public abstract class InternalArangoVertexCollection<A extends InternalArangoDB<
 		request = request(graph.db().name(), RequestType.PATCH, PATH_API_GHARIAL, graph.name(), VERTEX,
 			DocumentUtil.createDocumentHandle(name, key));
 		final VertexUpdateOptions params = (options != null ? options : new VertexUpdateOptions());
+		request.putHeaderParam(TRANSACTION_ID, params.getStreamTransactionId());
 		request.putQueryParam(ArangoRequestParam.KEEP_NULL, params.getKeepNull());
 		request.putQueryParam(ArangoRequestParam.WAIT_FOR_SYNC, params.getWaitForSync());
 		request.putHeaderParam(ArangoRequestParam.IF_MATCH, params.getIfMatch());
@@ -154,6 +160,7 @@ public abstract class InternalArangoVertexCollection<A extends InternalArangoDB<
 		final Request request = request(graph.db().name(), RequestType.DELETE, PATH_API_GHARIAL, graph.name(), VERTEX,
 			DocumentUtil.createDocumentHandle(name, key));
 		final VertexDeleteOptions params = (options != null ? options : new VertexDeleteOptions());
+		request.putHeaderParam(TRANSACTION_ID, params.getStreamTransactionId());
 		request.putQueryParam(ArangoRequestParam.WAIT_FOR_SYNC, params.getWaitForSync());
 		request.putHeaderParam(ArangoRequestParam.IF_MATCH, params.getIfMatch());
 		return request;