+++ src/main/java/com/arangodb/model/VertexUpdateOptions.java
@@ -30,6 +30,7 @@ public class VertexUpdateOptions {
 	private Boolean keepNull;
 	private Boolean waitForSync;
 	private String ifMatch;
+    private String streamTransactionId;
 
 	public VertexUpdateOptions() {
 		super();
@@ -79,4 +80,18 @@ public class VertexUpdateOptions {
 		this.ifMatch = ifMatch;
 		return this;
 	}
+
+    public String getStreamTransactionId() {
+        return streamTransactionId;
+    }
+
+    /**
+     * @param streamTransactionId If set, the operation will be executed within the transaction.
+     * @return options
+     * @since ArangoDB 3.5.1
+     */
+    public VertexUpdateOptions streamTransactionId(final String streamTransactionId) {
+        this.streamTransactionId = streamTransactionId;
+        return this;
+    }
 }